Add Info Row and format code
