+++ .eslintrc.json
@@ -14,6 +14,6 @@
   "plugins": ["react"],
   "rules": {
     "react/prop-types": 0,
-    "react/display-name":0
+    "react/display-name": 0
   }
 }
