+++ src/atoms/icon/Icon.module.css
