+++ src/atoms/input/Input.module.css
@@ -6,7 +6,7 @@ input {
 }
 
 .primary {
-  background-color: #2B2E3C;
+  background-color: #2b2e3c;
   color: #ffffff;
 }
