+++ src/molecules/feature/Features.module.css
@@ -1,0 +1,12 @@
+.nonBorderedItem div {
+  border: none;
+}
+
+.borderedItem div {
+  border-left: 1px solid black;
+}
+
+.borderedItem div:nth-child(1) {
+  border-left: none;
+}
+/*# sourceMappingURL=Features.module.css.map */
