+++ src/molecules/feature/Features.module.scss
@@ -1,10 +1,4 @@
-.FeaturesWrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 3rem;
-  min-height: 100%;
-}
+
 
 .nonBorderedItem div{
     border: none;
