--- src/organisms/home/Features.module.css
@@ -1,27 +1,0 @@
-.FeaturesWrapper {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  gap: 3rem;
-  min-height: 100%;
-}
-
-.nonBorderedItem div {
-  border: none;
-}
-
-.borderedItem div {
-  border-left: 1px solid black;
-}
-
-.borderedItem div:nth-child(1) {
-  border-left: none;
-}
-/*# sourceMappingURL=Features.module.css.map */
