+++ src/pages/Home/HomePage.js
@@ -1,54 +1,16 @@
 import React from 'react';
 import Row from '../../atoms/row/Row';
 import CTA from '../../organisms/cta/CTA';
-import Features from '../../organisms/home/Features';
+import FeatureHome1 from '../../organisms/home/FeatureHome1';
+import FeatureHome2 from '../../organisms/home/FeatureHome2';
 import TrialForm from '../../organisms/home/TrialForm';
 
 const HomePage = () => {
-  const features1 = [
-    {
-      icon: 'Guard.svg',
-      title: 'High security to protect from piracy',
-      desc: 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua invidunt ut labore.'
-    },
-    {
-      icon: 'Thumb.svg',
-      title: 'Premium quality performance',
-      desc: 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua invidunt ut labore.'
-    },
-    {
-      icon: 'Headset.svg',
-      title: 'Full time customer support - 24/7',
-      desc: 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua invidunt ut labore.'
-    }
-  ];
-  const features2 = [
-    {
-      icon: 'Gear.svg',
-      title: 'Advanced 256-bit encryption',
-      desc: 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua invidunt ut labore.'
-    },
-    {
-      icon: 'IconMessenger.svg',
-      title: 'Simple collaboration tools',
-      desc: 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua invidunt ut labore.'
-    },
-    {
-      icon: 'AI.svg',
-      title: 'Customizable AI features',
-      desc: 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua invidunt ut labore.'
-    }
-  ];
-  const featureBorder = true;
   return (
     <Row>
       <TrialForm />
-      <Features
-        features={features1}
-        title="We have the Best Solution for your Busines"
-        color="violet"
-      />
-      <Features features={features2} border={featureBorder} />
+      <FeatureHome1 />
+      <FeatureHome2 />
       <CTA />
     </Row>
   );
