+++ src/settings/variables.css
@@ -1,1 +1,1 @@
-/* No CSS *//*# sourceMappingURL=variables.css.map */
+/* No CSS */ /*# sourceMappingURL=variables.css.map */
